Updated and renamed LoadingTIle to Skeleton
