+++ es/components/Loader/index.js
@@ -2,4 +2,4 @@ export { default as Loader } from './Loader';
 export { default as CircleLoader } from './CircleLoader';
 export { default as BounceLoader } from './BounceLoader';
 export { default as HeartLoader } from './HeartLoader';
-export { default as LoadingTile } from './LoadingTile';
+export { default as Skeleton } from './Skeleton';
