+++ lib/components/Loader/Skeleton.js
@@ -35,11 +35,12 @@ var LoadingTile = (0, _styledComponents2.default)(_Box.Box)(_templateObject2, fu
 }, (0, _styledSystem.variant)({
   prop: 'ratio',
   variants: {
+    '4/5': { paddingTop: '125%' },
     '1/1': { paddingTop: '100%' },
     '3/2': { paddingTop: '66%' },
     '4/3': { paddingTop: '75%' },
-    text: { paddingTop: '16px' },
-    title: { paddingTop: '22px' }
+    text: { paddingTop: '20px' },
+    title: { paddingTop: '26px' }
   }
 }), loading);
