+++ lib/components/Loader/index.js
@@ -38,12 +38,12 @@ Object.defineProperty(exports, 'HeartLoader', {
   }
 });
 
-var _LoadingTile = require('./LoadingTile');
+var _Skeleton = require('./Skeleton');
 
-Object.defineProperty(exports, 'LoadingTile', {
+Object.defineProperty(exports, 'Skeleton', {
   enumerable: true,
   get: function get() {
-    return _interopRequireDefault(_LoadingTile).default;
+    return _interopRequireDefault(_Skeleton).default;
   }
 });
